Manual

Equation Learning

VandenHeuvel et al. (2022)

Paper results

Simulation studies

Manual

Structs

Bootstrapping

Basis Bootstrapping

Gaussian Processes

Optimisation

PDEs

Miscellaneous

Tutorial

PDE parameters

Bootstrapping parameters

Gaussian process parameters

Defining the functions

Fitting the model

Looking at the results

Structs

Bootstrapping

EquationLearning.bootstrap_gp — Function

bootstrap_gp(x::T1, t::T1, u::T1,
    T::Function, D::Function, D′::Function, R::Function,
    α₀::T1, β₀::T1, γ₀::T1, lowers::T1, uppers::T1;
    gp_setup::GP_Setup = GP_Setup(u),
    bootstrap_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u),
    optim_setup::Optim.Options = Optim.Options(),
    pde_setup::PDE_Setup = PDE_Setup(x),
    D_params = nothing, R_params = nothing, T_params = nothing, zvals = nothing, PDEkwargs...) where {T1<:AbstractVector}

Perform bootstrapping on the data (x, t, u) to learn the appropriate functional forms of (T, D, R) with uncertainty.

Arguments

x::T1: The spatial data.

t::T1: The temporal data.

u::T1: The density data.

T::Function: The delay function, given in the formT(t, α, T_params)`.

D::Function: The diffusion function, given in the form D(u, β, D_params).

D′::Function: The derivative of the diffusion function, given in the form D′(u, β, D_params).

R::Function: The reaction function, given in the form R(u, γ, R_params).

R′::Function: The derivative of the reaction function, given in the form R′(u, γ, R_params).

α₀::T1: Initial estimates of the delay parameters. Not actually used for anything other than ensuring the functions are specified correctly.

β₀::T1: Initial estimates of the diffusion parameters. Not actually used for anything other than ensuring the functions are specified correctly.

γ₀::T1: Initial estimates of the reaction parameters. Not actually used for anything other than ensuring the functions are specified correctly.

lowers::T1: Lower bounds to use for constructing the Latin hypersquare design, and for the constrained problem if constrained = true in bootstrap_setup.

uppers::T1: Upper bounds to use for constructing the Latin hypersquare design, and for the constrained problem if constrained = true in bootstrap_setup.

Keyword Arguments

gp_setup::GP_Setup = GP_Setup(u): Defines the setup for the Gaussian process. See also GP_Setup.

bootstart_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u): Defines some extra keyword arguments for the bootstrapping and optimisation process. See also Bootstrap_Setup.

optim_setup::Optim.Options = Optim_Options(): Defines some options when using Optim.optimize.

pde_setup::PDE_Setup = PDE_Setup(x): Defines some extra keyword arguments for the PDE solutions. See also PDE_Setup.

D_params = nothing: Extra known parameters for the diffusion function D.

R_params = nothing: Extra known parameters for the reaction function R.

T_params = nothing: Extra known parameters for the delay function T.

PDEkwargs...: Extra keyword arguments to use inside DifferentialEquations.solve.

Outputs

bgp: A BootResults structure. See BootResults.

Basis Bootstrapping

Gaussian Processes

Optimisation

PDEs

Miscellaneous