Manual

This section lists some of the docstrings for the functions defined from our module. All functions in this package are documented, though we only the important ones that are likely to be used by a user here. The remaining docstrings in this package are listed in the index section.

Structs

EquationLearning.GP_Setup — Type

GP_Setup

Setup for the Gaussian processes. See also fit_GP and bootstrap_gp.

Fields

ℓₓ::Vector{Float64}: A 2-vector giving the lower and upper bounds for the initial estimates of ℓₓ (defined on a log scale).

ℓₜ::Vector{Float64}: A 2-vector giving the lower and upper bounds for the initial estimates of ℓₜ (defined on a log scale).

σ::Vector{Float64}: A 2-vector giving the lower and upper bounds for the initial estimates of σ (defined on a log scale).

σₙ::Vector{Float64}: A 2-vector giving the lower and upper bounds for the initial estimates of σₙ (defined on a log scale).

GP_Restarts::Int: Number of times to restart the optimiser. See opt_restart!.

μ::Union{Missing, Vector{Float64}}: Either Nothing or a stored mean vector for the Gaussian process.

L::Union{Missing, LowerTriangular{Float64}}: Either Nothing or a stored Cholesky factor for the Gaussian process.

nugget::Float64: Nugget term to add to the covariance matrix to be symmetric positive definite. This nugget is adapted to the blocks of the matrix based on the derivatives, as described in our paper.

gp::Union{Missing, GPBase}: Either Nothing or a stored Gaussian process. See also fit_GP.

EquationLearning.Bootstrap_Setup — Type

struct Bootstrap_Setup

A struct defining some arguments for bootstrap_gp.

Fields

bootₓ::AbstractVector: The spatial grid for bootstrapping.

bootₜ::AbstractVector: The temporal grid for bootstrapping.

B::Int: Number of bootstrap samples.

τ::Tuple{Float64, Float64}: A tuple of the form (τ₁, τ₂) which gives the tolerance τ₁ for thresholding f and τ₂ for thresholding fₜ. See also data_thresholder.

Optim_Restarts::Int: Number of times to restart the optimiser for the nonlinear least squares problem. See also learn_equations!.

constrained::Bool: true if the optimisation problems should be constrained, and false otherwise.

obj_scale_GLS::Function: The function determining how the GLS loss function should be scaled.

obj_scale_PDE::Function: The function determining how the PDE loss function should be scaled.

init_weight::Float64: Weight factor for the initial condition for the GLS errors.

show_losses::Bool: true if the loss function should be printed to the REPL throughout the optimisation process, and false otherwise.

EquationLearning.PDE_Setup — Type

struct PDE_Setup

A struct defining some arguments for the PDEs in bootstrap_gp and boot_pde_solve.

Fields

meshPoints::AbstractVector: The spatial mesh to use for solving the PDEs involved when computing the "GLS" loss function, or for the PDEs in general.

LHS::Vector{Float64}: Vector defining the left-hand boundary conditions for the PDE. See also the definitions of (a₀, b₀, c₀) in sysdegeneral!.

RHS::Vector{Float64}: Vector defining the right-hand boundary conditions for the PDE. See also the definitions of (a₁, b₁, c₁) in sysdegeneral!.

finalTime::Float64: The final time to solve the PDE up to.

δt::Union{AbstractVector, Float64}: A number or a vector specifying the spacing between returned times for the solutions to the PDEs or specific times, respectively.

alg: Algorithm to use for solving the PDEs. If you want to let DifferentialEquations.jl select the algorithm automatically, specify alg = nothing. If automatic differentiation is being used in the ODE algorithm, then no Sundials algorithms can be used.

EquationLearning.BootResults — Type

struct BootResults

Structure for storing bootstrapping results. See bootstrap_gp.

Fields

delayBases: The estimated delay parameters. Each column corresponds to a single bootstrap iteration.

diffusionBases: The estimated diffusion parameters. Each column corresponds to a single bootstrap iteration.

reactionBases: The estimated reaction parameters. Each column corresponds to a single bootstrap iteration.

gp: The fitted Gaussian process. See fit_gp.

zvals: The simulated normal variables from N(0, I) used for drawing from the Gaussian process gp. See draw_gp!.

Xₛ: The test matrix for the bootstrapping grid data, given in the scale [0, 1].

Xₛⁿ: The unscaled form of Xₛ.

bootₓ: The spatial bootstrapping grid.

bootₜ: The temporal bootstrapping grid.

T: The delay function, given in the form T(t, α, T_params).

D: The diffusion function, given in the form D(u, β, D_params).

D′: The derivative of the diffusion function, given in the form D′(u, β, D_params).

R: The reaction function, given in the form R(u, γ, R_params).

R′: The derivative of the reaction function, given in the form R′(u, γ, R_params).

D_params: Parameters for the diffusion function.

R_params: Parameters for the reaction function.

T_params: Parameters for the delay function.

gp_setup: The GP setup used; see GP_Setup.

bootstrap_setup: The bootstrap setup used; see bootstrap_setup.

pde_setup: The PDE setup used; see pde_setup.

EquationLearning.BasisBootResults — Type

struct BasisBootResults

Structure for storing bootstrapping results from the basis approach. See basis_bootstrap_gp.

Fields

diffusionBases: The estimated diffusion parameters. Each column corresponds to a single bootstrap iteration.

reactionBases: The estimated reaction parameters. Each column corresponds to a single bootstrap iteration.

gp: The fitted Gaussian process. See fit_gp.

zvals: The simulated normal variables from N(0, I) used for drawing from the Gaussian process gp. See draw_gp!.

Xₛ: The test matrix for the bootstrapping grid data, given in the scale [0, 1].

Xₛⁿ: The unscaled form of Xₛ.

bootₓ: The spatial bootstrapping grid.

bootₜ: The temporal bootstrapping grid.

D: The diffusion basis functions, each given in the form φ(u, D_params).

D′: The derivative of the diffusion basis functions.

R: The reaction basis function, each given in the form ψ(u, R_params).

R′: The derivative of the reaction basis functions.

D_params: Parameters for the diffusion function.

R_params: Parameters for the reaction function.

gp_setup: The GP setup used; see GP_Setup.

bootstrap_setup: The bootstrap setup used; see bootstrap_setup.

pde_setup: The PDE setup used; see pde_setup.

EquationLearning.AllResults — Type

struct AllResults

Helpful structure for displaying results. This differs from e.g. BootResults since we also include PDE solutions.

Fields

pde_solutions::Array{Float64}: Solutions for the PDEs with sampled initial conditions.

AIC::Vector{Float64}: The AICs.

bgp::Union{BootResults, BasisBootResults}: Bootstrapping results.

delayCIs: Confidence intervals for the delay parameters.

diffusionCIs: Confidence intervals for the diffusion parameters.

reactionCIS: Confidence intervals for the reaction parameters.

delay_density::Figure: The density plots for the delay coefficients.

diffusion_density::Figure: The density plots for the diffusion coefficients.

reaction_density::Figure: The density plots for the reaction coefficients.

delay_curve::Figure: Plot for the learned delay curve.

diffusion_curve::Figure: Plot for the learned diffusion curve.

reaction_curve::Figure: Plot for the learned reaction curve.

pde_plot::Figure: Plot for the PDE solutions.

pde_error::Vector{Float64}: Confidence interval for the PDE errors.

Bootstrapping

EquationLearning.bootstrap_gp — Function

bootstrap_gp(x::T1, t::T1, u::T1,
    T::Function, D::Function, D′::Function, R::Function,
    α₀::T1, β₀::T1, γ₀::T1, lowers::T1, uppers::T1;
    gp_setup::GP_Setup = GP_Setup(u),
    bootstrap_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u),
    optim_setup::Optim.Options = Optim.Options(),
    pde_setup::PDE_Setup = PDE_Setup(x),
    D_params = nothing, R_params = nothing, T_params = nothing, zvals = nothing, PDEkwargs...) where {T1<:AbstractVector}

Perform bootstrapping on the data (x, t, u) to learn the appropriate functional forms of (T, D, R) with uncertainty.

Arguments

x::T1: The spatial data.

t::T1: The temporal data.

u::T1: The density data.

T::Function: The delay function, given in the formT(t, α, T_params)`.

D::Function: The diffusion function, given in the form D(u, β, D_params).

D′::Function: The derivative of the diffusion function, given in the form D′(u, β, D_params).

R::Function: The reaction function, given in the form R(u, γ, R_params).

R′::Function: The derivative of the reaction function, given in the form R′(u, γ, R_params).

α₀::T1: Initial estimates of the delay parameters. Not actually used for anything other than ensuring the functions are specified correctly.

β₀::T1: Initial estimates of the diffusion parameters. Not actually used for anything other than ensuring the functions are specified correctly.

γ₀::T1: Initial estimates of the reaction parameters. Not actually used for anything other than ensuring the functions are specified correctly.

lowers::T1: Lower bounds to use for constructing the Latin hypersquare design, and for the constrained problem if constrained = true in bootstrap_setup.

uppers::T1: Upper bounds to use for constructing the Latin hypersquare design, and for the constrained problem if constrained = true in bootstrap_setup.

Keyword Arguments

gp_setup::GP_Setup = GP_Setup(u): Defines the setup for the Gaussian process. See also GP_Setup.

bootstart_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u): Defines some extra keyword arguments for the bootstrapping and optimisation process. See also Bootstrap_Setup.

optim_setup::Optim.Options = Optim_Options(): Defines some options when using Optim.optimize.

pde_setup::PDE_Setup = PDE_Setup(x): Defines some extra keyword arguments for the PDE solutions. See also PDE_Setup.

D_params = nothing: Extra known parameters for the diffusion function D.

R_params = nothing: Extra known parameters for the reaction function R.

T_params = nothing: Extra known parameters for the delay function T.

PDEkwargs...: Extra keyword arguments to use inside DifferentialEquations.solve.

Outputs

bgp: A BootResults structure. See BootResults.

EquationLearning.basis_bootstrap_gp — Function

basis_bootstrap_gp(x::T1, t::T1, u::T1,
    D::Vector{Function}, D′::Vector{Function}, R::Vector{Function}, R′::Vector{Function};
    gp_setup::GP_Setup = GP_Setup(u),
    bootstrap_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u),
    optim_setup::Optim.Options = Optim.Options(),
    pde_setup::PDE_Setup = PDE_Setup(x),
    D_params = nothing, R_params = nothing, PDEkwargs...) where {T1<:AbstractVector}

Perform bootstrapping on the data (x, t, u) to learn the appropriate functional forms of (T, D, R) with uncertainty, using the basis function approach.

Arguments

x::T1: The spatial data.

t::T1: The temporal data.

u::T1: The density data.

D::Vector{Function}: The diffusion function, given in the form D(u, β, D_params).

D′::Vector{Function}: The derivative of the diffusion function, given in the form D′(u, β, D_params).

R::Vector{Function}: The reaction function, given in the form R(u, γ, R_params).

R′::Vector{Function}: The derivative of the reaction function, given in the form R′(u, γ, R_params).

Keyword Arguments

gp_setup::GP_Setup = GP_Setup(u): Defines the setup for the Gaussian process. See also GP_Setup.

bootstart_setup::Bootstrap_Setup = Bootstrap_Setup(x, t, u): Defines some extra keyword arguments for the bootstrapping and optimisation process. See also Bootstrap_Setup.

pde_setup::PDE_Setup = PDE_Setup(x): Defines some extra keyword arguments for the PDE solutions. See also PDE_Setup.

D_params = nothing: Extra known parameters for the diffusion function D.

R_params = nothing: Extra known parameters for the reaction function R.

PDEkwargs...: Extra keyword arguments to use inside DifferentialEquations.solve.

Outputs

bgp: A BasisBootResults structure. See BasisBootResults.

EquationLearning.update_results — Function

update_results(bgp, bgp_new, mechanism)

Updates the results in bgp with new results bgp_new for a re-estimated mechanism.

Gaussian Processes

EquationLearning.fit_GP — Function

fit_GP(x, t, u; <keyword arguments>)

Fits a Gaussian process with data (x, t) using the targets in u.

Arguments

x: The spatial data.

t: The temporal data.

u: The targets corresponding to (x, t).

Keyword Arguments

ℓₓ = log.([1e-4, 1.0]): A 2-vector giving the lower and upper bounds for the initial estimates of ℓₓ (defined on a log scale).

ℓₜ = log.([1e-4, 1.0]): A 2-vector giving the lower and upper bounds for the initial estimates of ℓₜ (defined on a log scale).

σ = log.([1e-1, 2std(u)]): A 2-vector giving the lower and upper bounds for the initial estimates of σ (defined on a log scale).

σₙ = log.([1e-5, 2std(u)]): A 2-vector giving the lower and upper bounds for the initial estimates of σₙ (defined on a log scale).

num_restarts = 50: Number of times to restart the optimiser. See opt_restart!.

Outputs

gp: The fitted Gaussian process.

fit_GP(x, t, u, setup::GP_Setup)

Method for calling fit_GP with a setup defined by GP_Setup.

EquationLearning.compute_joint_GP — Function

compute_joint_GP(gp::GPBase, X̃; nugget = 1e-10)

Computes the mean vector μ and Cholesky factor L such that LLᵀ = Σ, where

$\left[\begin{array}{c} \mathbf{f} \\ \frac{\partial\mathbf f}{\partial t} \\ \frac{\partial\mathbf f}{\partial x} \\\frac{\partial^2\mathbf f}{\partial x^2}\end{array}\right] \sim \mathcal N\left(\mathbf{\mu}, \mathbf{\Sigma}\right).$

Arguments

gp::GPBase: The fitted Gaussian process.

X̃: Test data for the Gaussian process.

Keyword Arguments

nugget = 1e-10: The term to add to the diagonals of the covariance matrix in case the matrix is not positive definite.

Outputs

μ: The mean vector.

L: The Cholesky factor of the covariance matrix.

Extended help

The covariance matrices are built without any attention to symmetry. The loops could be optimised by e.g. considering only the upper triangular components.

The covariance matrices are built using separate functions for the derivatives of the kernel function (currently only implemented for the squared exponential kernel). These functions are:

dkxⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²kxⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

dktⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

dktᵢ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²ktᵢtⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²ktᵢxⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d³tᵢxⱼ²(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

dkxᵢ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²kxᵢtⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²kxᵢxⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d³kxᵢxⱼ²(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d²kxᵢ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d³kxᵢ²tⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d³kxᵢ²xⱼ(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

d⁴kxᵢ²xⱼ²(x₁, t₁, x₂, t₂, ℓ₁, ℓ₂).

In these functions, the derivatives are evaluated at ([x₁; t₁], [x₂; t₂]) with length scales (ℓ₁, ℓ₂) for space and time, respectively. The functions are missing a multiplication by the kernel value cov(gp.kernel, [x₁; t₁], [x₂; t₂]) to allow for it to be done in place more easily. Note also that in some of these functions, powers are written as products, e.g. x³ = x*x*x, to allow for the @muladd macro to work (see https://github.com/SciML/MuladdMacro.jl). Since the functions are quite small, we use @inline on each function to encourage the compiler to inline the function in the LLVM.

You may get a StackOverflowError from this function. This may be related to https://github.com/JuliaLang/julia/issues/43242, in which case you can set the number of BLAS threads to 1:

julia> ccall((:openblas_get_num_threads64_, Base.libblas_name), Cint, ()) # In case you need to remember how many threads you used prior
julia> LinearAlgebra.BLAS.set_num_threads(1)

EquationLearning.precompute_gp_mean — Function

precompute_gp_mean(x, t, u, ℓₓ, ℓₜ, σ, σₙ, nugget, num_restarts, bootstrap_setup::Bootstrap_Setup)

Computes the Gaussian process and corresponding mean vector and Cholesky factor for a joint Gaussian process defined by the data (x, t, u). See also compute_joint_GP and bootstrap_gp.

Arguments

x: The spatial data.

t: The temporal data.

u: The targets corresponding to (x, t).

ℓₓ: A 2-vector giving the lower and upper bounds for the initial estimates of ℓₓ (defined on a log scale).

ℓₜ: A 2-vector giving the lower and upper bounds for the initial estimates of ℓₜ (defined on a log scale).

σ: A 2-vector giving the lower and upper bounds for the initial estimates of σ (defined on a log scale).

σₙ: A 2-vector giving the lower and upper bounds for the initial estimates of σₙ (defined on a log scale).

nugget: The term to add to the diagonals of the covariance matrix in case the matrix is not positive definite.

num_restarts = 50: Number of times to restart the optimiser. See opt_restart!.

bootstrap_setup::Bootstrap_Setup: A Bootstrap_Setup struct for defining the bootstrapping grid.

Outputs

gp_setup: The fitted Gaussian process.

μ: The mean vector for the joint Gaussian proces.

L: The Cholesky factor for the joint Gaussian process.

`

PDEs

EquationLearning.boot_pde_solve — Function

boot_pde_solve(bgp::BootResults, x_pde, t_pde, u_pde; ICType = "data")

Solve the PDEs corresponding to the bootstrap iterates in bgp obtained from bootstrap_gp.

Arguments

bgp::BootResults: A BootResults struct containing the results from bootstrap_gp.

x_pde: The spatial data to use for obtaining the initial condition.

t_pde: The temporal data to use for obtaining the initial condition.

u_pde: The density data to use for obtaining the initial condition.

Keyword Arguments

ICType = "data": The type of initial condition to use. Should be either "data" or "gp".

Outputs

solns_all: The solutions to the PDEs over the mesh points at each time value.

Note

The _pde subscript is used to indicate that these data need not be the same as the (x, t, u) used in bootstrap_gp, for example. For example, we may have 3 replicates of some data which we would easily use in bootstrap_gp, but for the PDE we would need to average these together for obtaining the solutions.

boot_pde_solve(bgp::BasisBootResults, x_pde, t_pde, u_pde, ICType = "data")

Solve the PDEs corresponding to the bootstrap iterates in bgp obtained from basis_bootstrap_gp.

Arguments

bgp::BasisBootResults: A BasisBootResults struct containing the results from basis_bootstrap_gp.

x_pde: The spatial data to use for obtaining the initial condition.

t_pde: The temporal data to use for obtaining the initial condition.

u_pde: The density data to use for obtaining the initial condition.

Keyword Arguments

ICType = "data": The type of initial condition to use. Should be either "data" or "gp".

Outputs

solns_all: The solutions to the PDEs over the mesh points at each time value.

Note

The _pde subscript is used to indicate that these data need not be the same as the (x, t, u) used in bootstrap_gp, for example. For example, we may have 3 replicates of some data which we would easily use in bootstrap_gp, but for the PDE we would need to average these together for obtaining the solutions.

EquationLearning.error_comp — Function

error_comp(bgp, solns_all, x, t, u; level = 0.05)

Computes the error between solutions to a PDE compared to data (x, t, u) from the bootstrapping procedure. The error measure used is median(100 * (absolute errors) / max(eps(Float64), u)).

Arguments

bgp: Bootstrapping results.

solns_all: PDE solutions.

x: Spatial data.

t: Temporal data.

u: Density data.

Keyword Arguments

level = 0.05: Level for the confidence interval.

compute_mean = false: Whether to only report the mean.

Outputs

err_CI: 100(1-level)% confidence interval for the error.

Plotting

EquationLearning.density_values — Function

density_values(bgp::BootResults; <keyword arguments>)

Computes the densities for the bootstrapping results in bgp.

Arguments

bgp: A BootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

delay_scales = nothing: Values that multiply the individual delay parameters.

diffusion_scales = nothing: Values that multiply the individual diffusion parameters.

reaction_scales = nothing: Values that multiply the individual reaction parameters.

Outputs

trv: Delay densities.

dr: Diffusion densities.

rr: Reaction densities.

tt: Number of delay parameters.

d: Number of diffusion parameters.

r: Number of reaction parameters.

delayCIs: Confidence intervals for the delay parameters.

diffusionCIs: Confidence intervals for the diffusion parameters.

reactionCIS: Confidence intervals for the reaction parameters.

density_values(bgp::BasisBootResults; <keyword arguments>)

Computes the densities for the bootstrapping results in bgp from a basis function approach.

Arguments

bgp: A BasisBootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

diffusion_scales = nothing: Values that multiply the individual diffusion parameters.

reaction_scales = nothing: Values that multiply the individual reaction parameters.

Outputs

dr: Diffusion densities.

rr: Reaction densities.

d: Number of diffusion parameters.

r: Number of reaction parameters.

delayCIs: Confidence intervals for the delay parameters.

diffusionCIs: Confidence intervals for the diffusion parameters.

reactionCIS: Confidence intervals for the reaction parameters.

EquationLearning.density_results — Function

density_results(bgp::BootResults; <keyword arguments>)

Plots the densities for the bootstrapping results in bgp.

Arguments

bgp: A BootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

delay_scales = nothing: Values that multiply the individual delay parameters.

diffusion_scales = nothing: Values that multiply the individual diffusion parameters.

reaction_scales = nothing: Values that multiply the individual reaction parameters.

diffusion_resolution = (800, 800): Resolution for the diffusion figure.

reaction_resolution = (800, 800): Resolution for the reaction figure.

delay_resolution = (800, 800): Resolution for the delay fgure.

Outputs

delayDensityFigure: A figure of plots containing a density plot for each delay parameter.

diffusionDensityFigure: A figure of plots containing a density plot for each diffusion parameter.

reactionDensityFigure: A figure of plots containing a density plot for each reaction parameter.

density_results(bgp::BasisBootResults; <keyword arguments>)

Plots the densities for the bootstrapping results in bgp with the basis function approach.

Arguments

bgp: A BasisBootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

diffusion_scales = nothing: Values that multiply the individual diffusion parameters.

reaction_scales = nothing: Values that multiply the individual reaction parameters.

diffusion_resolution = (800, 800): Resolution for the diffusion figure.

reaction_resolution = (800, 800): Resolution for the reaction figure.

Outputs

diffusionDensityFigure: A figure of plots containing a density plot for each diffusion parameter.

reactionDensityFigure: A figure of plots containing a density plot for each reaction parameter.

EquationLearning.curve_values — Function

curve_values(bgp::BootResults; <keyword arguments>)

Computes values for plotting the learned functional forms along with confidence intervals for the bootstrapping results in bgp.

Arguments

bgp: A BootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

x_scale = 1.0: Value used for scaling the spatial data (and all other length units, e.g. for diffusion).

t_scale = 1.0: Value used for scaling the temporal data (and all other time units, e.g. for reaction).

Outputs

Tu_vals: Ribbon features for the delay functions.

Du_vals: Ribbon features for the diffusion functions.

Ru_vals: Ribbon features for the reaction functions.

u_vals: The density values used for computing the functions.

t_vals: The time values used for computing the functions.

curve_values(bgp::BasisBootResults; <keyword arguments>)

Computes values for plotting the learned functional forms along with confidence intervals for the bootstrapping results in bgp from the basis function approach.

Arguments

bgp: A BasisBootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

x_scale = 1.0: Value used for scaling the spatial data (and all other length units, e.g. for diffusion).

t_scale = 1.0: Value used for scaling the temporal data (and all other time units, e.g. for reaction).

Outputs

Du_vals: Ribbon features for the diffusion functions.

Ru_vals: Ribbon features for the reaction functions.

u_vals: The density values used for computing the functions.

t_vals: The time values used for computing the functions.

EquationLearning.curve_results — Function

curve_results(bgp::BootResults; <keyword arguments>)

Plots the learned functional forms along with confidence intervals for the bootstrapping results in bgp.

Arguments

bgp: A BootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

x_scale = 1.0: Value used for scaling the spatial data (and all other length units, e.g. for diffusion).

t_scale = 1.0: Value used for scaling the temporal data (and all other time units, e.g. for reaction).

Outputs

delayCurvePlots: A plot containing the learned functional form for the delay function, along with an uncertainty ribbon.

diffusionCurvePlots: A plot containing the learned functional form for the diffusion function, along with an uncertainty ribbon.

reactionCurvePlots: A plot containing the learned functional form for the reaction function, along with an uncertainty ribbon.

curve_results(bgp::BasisBootResults; <keyword arguments>)

Plots the learned functional forms along with confidence intervals for the bootstrapping results in bgp from the basis function approach.

Arguments

bgp: A BasisBootResults object which contains the results for the bootstrapping.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

x_scale = 1.0: Value used for scaling the spatial data (and all other length units, e.g. for diffusion).

t_scale = 1.0: Value used for scaling the temporal data (and all other time units, e.g. for reaction).

Outputs

diffusionCurvePlots: A plot containing the learned functional form for the diffusion function, along with an uncertainty ribbon.

reactionCurvePlots: A plot containing the learned functional form for the reaction function, along with an uncertainty ribbon.

EquationLearning.pde_values — Function

function pde_values(solns_all, bgp::Union{BootResults, BasisBootResults}; level = 0.05)

Plots the solutions to the PDEs corresponding to the bootstrap samples of bootstrap_gp using the computed solutins in boot_pde_solve.

Arguments

solns_all: The solutions to the PDEs for each bootstrap iteration.

bgp::Union{BootResults, BasisBootResults}: Bootstrapping results.

Keyword Arguments

level = 0.05: The significance level for computing the credible intervals for the parameter values.

Outputs

soln_vals_mean: Mean solution values.

soln_vals_lowers: Lower confidence interval for solution values.

soln_vals_upper: Upper confidence interval for solution values.

EquationLearning.pde_results — Function

pde_results(x_pde, t_pde u_pde, solns_all, bgp::Union{BootResults, BasisBootResults}; <keyword arguments>)

Plots the solutions to the PDEs corresponding to the bootstrap samples of bootstrap_gp using the computed solutins in boot_pde_solve.

Arguments

x_pde: The spatial data to plot as points.

t_pde: The temporal data to plot as points.

u_pde: The density data to plot as points.

solns_all: The solutions to the PDEs for each bootstrap iteration.

bgp::Union{BootResults, BasisBootResults}: A BootResults or BasisBootResults struct containing the results from bootstrap_gp or basis_bootstrap_gp, respectively.

Keyword Arguments

colors = [:black, :blue, :red, :magenta, :green]: A list of colors for colouring the solutions at each time.

level = 0.05: The significance level for computing the credible intervals for the parameter values.

fontsize = 23: Font size for the plots (to be used in plot_aes!).

x_scale = 1.0: Value used for scaling the spatial data (and all other length units, e.g. for diffusion).

t_scale = 1.0: Value used for scaling the temporal data (and all other time units, e.g. for reaction).`

Outputs

pdeSolutionPlots_BGP: The plot of the PDE solutions.

EquationLearning.delay_product — Function

delay_product(bgp, t; type = "diffusion", level = 0.05, x_scale = 1.0, t_scale = 1.0)

Computes the curve T(t; α)D(u; β) (if type = "diffusion") or T(t; β)R(u; γ) (if type = "reaction") with uncertainty for the results in bgp, at the time t. The significance level is determined by level (0.05 default) and the spatial and temporal data are scaled by x_scale (1.0 default) and t_scale (1.0 default), respectively.

Data Generation

EquationLearning.generate_data — Function

generate_data(x₀, u₀, T, D, R, α, β, γ, δt, finalTime; <keyword arguments>)

Generate synthetic data from values (x₀, t₀) at t = 0 with exact mechanisms T, D, and R for delay, diffusion, and reaction, respectively.

Arguments

x₀: The spatial mesh for points at t = 0.

u₀: The density values at t = 0 corresponding to the points in x₀.

T: A delay function of the form T(t, α, T_params).

D: A diffusion function of the form D(u, β, D_params).

R: A reaction function of the form R(u, γ, R_params).

D′: The derivative of the diffusion function, given in the form D′(u, β, D_params).

R′: The reaction function, given in the form R(u, γ, R_params).

α: Exact values for the delay parameters.

β: Exact values for the diffusion parameters.

γ: Exact values for the reaction parameters.

δt: Times to save the solution to the differential equations at for generating the data.

finalTime: The time to solve the differential equations up to.

Keyword Arguments

N = 1000: The number of mesh points to use.

LHS = [0.0, 1.0, 0.0]: Vector defining the left-hand boundary conditions for the PDE. See also the definitions of (a₀, b₀, c₀) in sysdegeneral!.

RHS = [0.0, -1.0, 0.0]: Vector defining the right-hand boundary conditions for the PDE. See also the definitions of (a₁, b₁, c₁) in sysdegeneral!.

alg = nothing: The algorithm to use for solving the differential equations.

N_thin = 100: The number of points to take from the solution at each time.

num_restarts = 50: The number of times to restart the optimiser for fitting the Gaussian process to the data.

D_params: Additional known parameters for the diffusion function.

R_params: Additional known parameters for the reaction function.

T_params: Additional known parameters for the delay function.

Model Comparison

EquationLearning.AIC — Function

AIC(bgp::Union{BootResults, BasisBootResults}; correct = true)

Computes all the AIC values for the results in bgp. A small-sample size correction is used if correct = true. The formulas used are given in Eq. 6 (uncorrected) or Eq. 17 (corrected) of Banks and Joyner (2017) [https://doi.org/10.1016/j.aml.2017.05.005].

EquationLearning.compare_AICs — Function

compare_AICs(AICs::Float64...)

Compares the AIC values in AICs.... See also classify_Δᵢ and AIC.

compare_AICs(AICs::Vector{Float64}...)

Compares many AICs by comparing entry-wise. The results are averaged. Assumes that each AIC is of equal length, otherwise computes only up to the minimum length. The returned value is such that the (i, j) entry gives the proportion of times that model i has been given interpretation j from classify_Δᵢ.

compare_AICs(x, t, u, models::Union{BootResults, BasisBootResults}...; correct = true)

Compare several bootstrapped models using AIC.